pyproject.toml
[project]
name = "nutrirag"
version = "0.1.0"
description = "AI Agent Nutrition RAG"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-groq>=0.2.0",
    "langchain-huggingface>=0.1.0",
    "langchain-qdrant>=0.1.4",
    "qdrant-client>=1.11.0",
    "sentence-transformers>=3.0.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",
    "pdfplumber>=0.11.0",
    "streamlit>=1.38.0",
    "llama-parse>=0.6.88",
    "ragas>=0.4.2",
    "datasets>=4.4.2",
    "pypdf>=6.6.0",
    "rapidfuzz>=3.14.3",
    "langchain-google-genai>=4.0.0",
]
[dependency-groups]
dev = [
    "ipykernel>=6.29.0",
]
